// makrom-cliente/src/app/interfaces/cliente.ts
export interface Cliente{
    name: string;
    email: string;
    phone: string;
    mobilePhone: string;
    cpfCnpj: string;
    postalCode: string;
    address: string;
    addressNumber: string;
    complement: string;
    province: string;
    city: string;
    state: string;
    externalReference: string;
    notificationDisabled: boolean;
    additionalEmails: string;
    municipalInscription: string;
    stateInscription: string;
    observations: string;
}