// makrom-cliente/src/app/interfaces/usuario.ts
export interface Usuario {
    id: number;
    name: string;
    
    bday: any;
    genre: string;
    cpfcnpj: number;
    phonenumber: number;
    homenumber: number;
    cep: number;
    address: string;
    number: number;
    city: string;
    nhood: string;
    state: string;
    province: string;
        email: string;
        usuario: string;
        senha: string;   
}